requirements.txt
django
djangorestframework
psycopg2-binary
cryptography
redis
rncryptor